calculateAbsencePercentage() + AbsenceCalculator.java
 calculateAbsencePercentage() in LogicManager
 calculateAbsence() in AbsenceCalculator

need to change the Student class back to using IntegerProperty for absence field.

+++ src/attendance/gui/controller/TeacherStudentAttendanceController.java
@@ -37,7 +37,7 @@ public class TeacherStudentAttendanceController implements Initializable {
     @FXML
     private TableColumn<Student, String> studentName;
     @FXML
-    private TableColumn<Student, Double> absence;
+    private TableColumn<Student, Integer> absence;
 
     private StudentModel studentModel;
     private CourseModel courseModel;
